refactor: JWTUtil 공통 모듈로 위치 변경
